Refactor/67 problem objective modify entity obo step separate and logic modify (#69)
* [BE] feat:(#51) 객관식 문제 업데이트 위한 작업중 obostep과 관련된 (entity) 문제가 찾아져 수정을 위한 임시 커밋

* [BE] feat:(#51) 객관식 문제 수정, 아주 간단한 일부 영역에서 수정이 되는것을 확인, 추가로 로직에 있어서 변경이 필요하며(null 문제) 동시에 obo step 분리 또한 필요함

* [BE] feat:(#51) problem 엔티티 수정 -1.  updatedAt의 updateable = false 제거, 2. updatedAt getter 추가, 3. 데이터 수정을 위한, 메서드 추가

* develop 변경사항(obo step entitiy) 반영 위해 merge 준비하기 위해 사전 commit

* [BE] refactor:(#67)객관식문제에 대해서, OboStep을 분리하고, 이에대한 생성 수정 삭제 로직을 수정함

+++ src/main/java/net/diveon/backend/domain/problem/service/ProblemDeleteService.java
@@ -1,12 +1,13 @@
 package net.diveon.backend.domain.problem.service;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import net.diveon.backend.domain.problem.dto.response.ProblemDeleteResponse;
-import net.diveon.backend.domain.problem.entity.Problem;
-import net.diveon.backend.domain.problem.entity.ProblemObjective;
+import net.diveon.backend.domain.problem.repository.OboStepRepository;
 import net.diveon.backend.domain.problem.repository.ProblemObjectiveRepository;
 import net.diveon.backend.domain.problem.repository.ProblemRepository;
+import net.diveon.backend.global.exception.ProblemNotFoundException;
 
 
 /**
@@ -36,19 +37,44 @@ public class ProblemDeleteService {
 
     private final ProblemRepository problemRepository;
     private final ProblemObjectiveRepository problemObjectiveRepository;
+    private final OboStepRepository oboStepRepository;
 
     
     public ProblemDeleteService(ProblemRepository problemRepository,
-        ProblemObjectiveRepository problemObjectiveRepository){
+        ProblemObjectiveRepository problemObjectiveRepository,
+        OboStepRepository oboStepRepository){
         this.problemRepository = problemRepository;
         this.problemObjectiveRepository = problemObjectiveRepository;
+        this.oboStepRepository = oboStepRepository;
     }
 
 
-    public ProblemDeleteResponse deleteProblemObjective(String userId, Long prodId){
-        Problem problem = problemRepository.findById(prodId).orElseThrow();
-        ProblemObjective problemObjective = problemObjectiveRepository.findById(prodId).orElseThrow();
+    @Transactional
+    public ProblemDeleteResponse deleteProblemObjective(String userId, long prodId){
+
+        // // 아래 코드는 검증을 좀더 깔끔한 방식으로 할 수 있다면 좋을듯. 
+        // problemRepository.findById(prodId).orElseThrow(ProblemNotFoundException::new);
+        // problemObjectiveRepository.findById(prodId).orElseThrow(ProblemNotFoundException::new);
+
+
+        if(!problemRepository.existsById(prodId)){
+            throw new ProblemNotFoundException("문제 전체 중" + prodId +"번에 해당하는 문제 자체가 존재하지 않습니다.");
+        }
+        if(!problemObjectiveRepository.existsById(prodId)){
+            throw new ProblemNotFoundException("객관식 문제 중" + prodId +"번에 해당하는 문제 자체가 존재하지 않습니다.");
+        }
         
+        //아래는 다른 방식으로, 예외시 기본 생성자 말고, 다른 생성자를 사용할 수 있는 방법.
+        // Problem problem = problemRepository.findById(prodId)
+        //     .orElseThrow(() -> new ProblemNotFoundException("ID " + prodId + "번에 해당하는 문제를 찾을 수 없습니다."));
+
+        // 널값에 대항할 수 있는 로직이 필요하다. 만약 long 이 아니라 Long 을 쓴다면.
+        // Problem problem = problemRepository.findById(prodId).orElseThrow();
+        // List<OboStep> oboSteps = oboStepRepository.findByProblemOrderByStepAsc(problem);
+        // if (!oboSteps.isEmpty()) {
+        //     oboStepRepository.deleteAll(oboSteps);
+        // }
+        oboStepRepository.deleteByProblem_Id(prodId);
         problemObjectiveRepository.deleteById(prodId);
         problemRepository.deleteById(prodId);
